/api/fields implemented

+++ CodeKickFc/src/main/java/com/codecool/codekickfc/dao/model/FootballPitch.java
@@ -1,3 +1,3 @@
 package com.codecool.codekickfc.dao.model;
 
-public record FootballPitch(int Id, String name, int maxPlayers, int price, String address) { }
+public record FootballPitch(int Id, String name, int maxPlayers, double price, String address) { }
